Update layout metadata and language settings
Modified the RootLayout component in layout.tsx to update the application title to "Comunidad Paraguay Tech" and the description to "Proyectando ideas, creando comunidad y ayudando a crecer". Additionally, changed the HTML language attribute from "en" to "es" to reflect the application's focus on the Spanish-speaking community. These changes enhance the application's branding and localization efforts.